+++ gitopscli/git/__init__.py
@@ -1,0 +1,1 @@
+from .git_factory import create_git
